+++ src/components/Films/Card/Card.scss
@@ -27,7 +27,7 @@
 
     &_active {
       background-image: url('./likeActive.svg');
-      cursor: default;
+      cursor: default !important;
     }
   }
