+++ src/redux/formSlice.ts
@@ -1,19 +1,12 @@
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 import { RootState } from './store';
-
-interface IGet {
-  name: string;
-  email: string;
-}
-
-interface IResponse extends IGet {
-  password?: string;
-}
-
-type TReg = {
-  email: string;
-  password?: string;
-};
+import {
+  getProfile,
+  IGet,
+  login,
+  register,
+  signout,
+} from './thunks/formThunks';
 
 type TStates = {
   isLogged: boolean;
@@ -21,71 +14,6 @@ type TStates = {
   pname: string;
   pemail: string;
 };
-
-export const login = createAsyncThunk<any, TReg>(
-  'login',
-  async ({ email, password }) => {
-    const res = await fetch('http://localhost:3001/signin', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include',
-      body: JSON.stringify({ email, password }),
-    });
-    const data = await res.json();
-    if (res.ok) {
-      console.log(res);
-      return data;
-    }
-    return Promise.reject('ошибка');
-  }
-);
-
-export const signout = createAsyncThunk<any>('/signout', async () => {
-  const res = await fetch('http://localhost:3001/signout', {
-    headers: { 'Content-Type': 'application/json' },
-    credentials: 'include',
-  });
-  if (res.ok) {
-    const data = await res.json();
-    console.log(data);
-    return data;
-  }
-  return Promise.reject('не получилось почистить куку');
-});
-
-export const register = createAsyncThunk<IGet, IResponse>(
-  'register',
-  async ({ name, email, password }: IResponse) => {
-    try {
-      const res = await fetch('http://localhost:3001/signup', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        credentials: 'include',
-        body: JSON.stringify({ name, email, password }),
-      });
-      if (res.ok) {
-        const data = await res.json();
-        console.log(data);
-        return data;
-      }
-      return Promise.reject('ошибочка');
-    } catch (error) {
-      console.log(error);
-    }
-  }
-);
-
-export const getProfile = createAsyncThunk<any>('/me', async () => {
-  const res = await fetch('http://localhost:3001/users/me', {
-    headers: { 'Content-Type': 'application/json' },
-    credentials: 'include',
-  });
-  if (res.ok) {
-    const data = await res.json();
-    return data;
-  }
-  return Promise.reject('ошибка получения профиля');
-});
 
 const initialState: TStates = {
   isLogged: false,
@@ -107,7 +35,7 @@ const formSlice = createSlice({
       state.isLogged = true;
     });
     builder.addCase(login.rejected, (state) => {
-      console.log('ошибка')
+      console.log('ошибка');
       state.isLogged = false;
     });
     builder.addCase(register.rejected, (state) => {
